# docs/swagger.yaml
basePath: /api
definitions:
  handlers.StringMap:
    additionalProperties:
      type: string
    type: object
  handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
  models.Course:
    properties:
      duration:
        type: string
      id:
        type: integer
    type: object
  models.Drug:
    properties:
      expiration_date:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is an API for managing users, drugs, courses etc.
  title: Medication API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StringMap'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Login user
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register user and returns id
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: register user
      tags:
      - auth
  /info/drug:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Drug'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Получить все препараты
      tags:
      - info
  /info/drug/{name}:
    get:
      parameters:
      - description: Название препарата
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Drug'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Получить препарат по названию
      tags:
      - info
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Получить всех пользователей
      tags:
      - users
  /users/{userid}:
    delete:
      parameters:
      - description: ID пользователя
        in: path
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Удалить пользователя
      tags:
      - users
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Получить пользователя по ID
      tags:
      - users
  /users/{userid}/{drugid}:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID пользователя
        in: path
        name: userid
        required: true
        type: integer
      - description: ID препарата
        in: path
        name: drugid
        required: true
        type: integer
      - description: Информация о препарате
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Drug'
      produces:
      - application/json
      responses:
        "200":
          description: Drug added
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Добавить препарат пользователю
      tags:
      - users
  /users/{userid}/drugs:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Drug'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Получить препараты пользователя
      tags:
      - users
  /users/{userid}/drugs/{drugid}:
    delete:
      parameters:
      - description: ID пользователя
        in: path
        name: userid
        required: true
        type: integer
      - description: ID препарата
        in: path
        name: drugid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Drug deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Удалить препарат пользователя
      tags:
      - users
  /users/{userid}/drugs/{drugid}/course:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: userid
        required: true
        type: integer
      - description: ID препарата
        in: path
        name: drugid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Получить курс по препарату
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID пользователя
        in: path
        name: userid
        required: true
        type: integer
      - description: ID препарата
        in: path
        name: drugid
        required: true
        type: integer
      - description: Новый курс
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Course'
      produces:
      - application/json
      responses:
        "200":
          description: Course edited
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Редактировать курс препарата
      tags:
      - users
swagger: "2.0"